Add support for HTTP 1.1 in StowRS requests
Brings back old version of StowRS that was removed when we upgraded to HTTP 2.0 to improve error handling and sets it as the default.

This fixes flakiness issues with HTTP 2.0 in the Healthcare API.

Tested by running with the flag on/off and confirmed that the default has no errors whereas turning this flag on has about 20% error rate when running adapter locally on my laptop.